+++ projects/js-packages/publicize-components/src/hooks/use-publicize-config/index.js
@@ -1,11 +1,14 @@
+import { useConnection } from '@automattic/jetpack-connection';
 import {
 	getJetpackExtensionAvailability,
 	isUpgradable,
 	getJetpackData,
 	getSiteFragment,
+	isSimpleSite,
 } from '@automattic/jetpack-shared-extension-utils';
 import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
+import { store as socialStore } from '../../social-store';
 import { usePostMeta } from '../use-post-meta';
 
 const republicizeFeatureName = 'republicize';
@@ -25,6 +28,7 @@ export default function usePublicizeConfig() {
 		getJetpackExtensionAvailability( republicizeFeatureName )?.available || isShareLimitEnabled;
 	const isPostPublished = useSelect( select => select( editorStore ).isCurrentPostPublished(), [] );
 	const currentPostType = useSelect( select => select( editorStore ).getCurrentPostType(), [] );
+	const { isUserConnected } = useConnection();
 
 	const connectionsRootUrl =
 		getJetpackData()?.social?.publicizeConnectionsUrl ??
@@ -105,6 +109,10 @@ export default function usePublicizeConfig() {
 	 */
 	const isJetpackSocialNote = 'jetpack-social-note' === currentPostType;
 
+	const needsUserConnection = ! isUserConnected && ! isSimpleSite();
+
+	const userConnectionUrl = useSelect( select => select( socialStore ).userConnectionUrl(), [] );
+
 	return {
 		isPublicizeEnabledMeta,
 		isPublicizeEnabled,
@@ -127,5 +135,7 @@ export default function usePublicizeConfig() {
 		isAutoConversionEnabled,
 		jetpackSharingSettingsUrl: getJetpackData()?.social?.jetpackSharingSettingsUrl,
 		isJetpackSocialNote,
+		needsUserConnection,
+		userConnectionUrl,
 	};
 }