+++ projects/js-packages/publicize-components/src/social-store/types.ts
@@ -7,10 +7,13 @@ export type SharesData = {
 	shared_posts_count: number;
 };
 
+export type ConnectionStatus = 'ok' | 'must_reauth' | 'invalid' | 'broken' | 'refresh-failed';
+
 export type Connection = {
 	id: string;
 	service_name: string;
 	display_name: string;
+	external_display?: string;
 	username: string;
 	enabled: boolean;
 	done: boolean;
@@ -21,10 +24,20 @@ export type Connection = {
 	can_disconnect: boolean;
 	profile_picture: string;
 	profile_link: string;
+	status: ConnectionStatus;
+};
+
+export type ConnectionService = {
+	ID: string;
+	label: string;
+	type: 'publicize' | 'other';
+	description: string;
+	connect_URL: string;
 };
 
 export type ConnectionData = {
 	connections: Connection[];
+	deletingConnections?: Array< number | string >;
 };
 
 export type JetpackSettings = {