+++ projects/packages/google-analytics/tests/php/bootstrap.php
@@ -9,3 +9,8 @@
  * Include the composer autoloader.
  */
 require_once __DIR__ . '/../../vendor/autoload.php';
+
+define( 'WP_DEBUG', true );
+
+\WorDBless\Load::load();
+require_once ABSPATH . WPINC . '/class-IXR.php';