+++ projects/packages/publicize/src/class-publicize.php
@@ -241,7 +241,7 @@ class Publicize extends Publicize_Base {
 	/**
 	 * Get all connections for a specific user.
 	 *
-	 * @return array|false
+	 * @return array
 	 */
 	public function get_all_connections_for_user() {
 		$connections = $this->get_all_connections();
@@ -259,7 +259,7 @@ class Publicize extends Publicize_Base {
 								array(
 									'service_name'   => $service_name,
 									'connection_id'  => $connection['connection_data']['id'],
-									'can_disconnect' => current_user_can( 'manage_options' ) || get_current_user_id() === $user_id,
+									'can_disconnect' => current_user_can( 'edit_others_posts' ) || get_current_user_id() === $user_id,
 									'profile_link'   => $this->get_profile_link( $service_name, $connection ),
 								)
 							);
@@ -269,11 +269,8 @@ class Publicize extends Publicize_Base {
 					}
 				}
 			}
-
-			return $connections_to_return;
 		}
-
-		return false;
+		return $connections_to_return;
 	}
 
 	/**